+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Added `validationRules` for `Form`
 - Fixed: missed selection state of `RadioButtonGroup` in non-controlled mode with form context.
 - Compatibility with IE11
+- Fixed: component wrapped with `withValidation` hoc didn't display an error
 
 ## 0.5.0